requirements.txt
coverage
pre-commit
pytest
requests
more-itertools
google-cloud-storage
